Rename useVocabularyResolver to useVocabulary
